Add language around cyclic nesting
After some back and forth the proposal was updated to include some Goals to better address the intent behind this approach.

This first iteration of the feature will cater for applications to be used in conjunction with `imgpkg` that know how to read the output format provided. In a later iteration, a better Human UX might be added to help visualize the structure that each bundle in the disk has.